requirements.txt
streamlit==1.45.1
pandas==2.2.3
numpy==2.2.6
joblib==1.4.2
Pillow==11.0.0
